Enable RS485 mode for USB_SERIAL_XR_RS485
The driver never supported switching to RS485 mode. Following steps must
be taken:
1. Change flow control register to set Half/Full duplex
2. Enable GPIO5 as RTS and keep TX HI

New 'mode' module parameter is introduced to list ports that have to be
switched to RS485. The format is following
mode=<port index><f|h>,...  where f - full duplex, h - half duplex.
If the port is not specified it is switched to RS232 mode

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>